Moved Anchor to AppFrame
